Fixed compiler warnings

+++ displayutils.cpp
@@ -59,7 +59,7 @@ namespace utils
     {
         QList<QVariant> list = value.toList();
         QString imploded;
-        for (size_t i = 0; i < list.size(); i++) {
+        for (int i = 0; i < list.size(); i++) {
             imploded += list[i].toString();
             if (i < list.size() - 1)
                 imploded += ", ";
